pyproject.toml
[build-system]
requires = [
  "setuptools >= 61.0",
  "setuptools_scm[toml] >= 8",
]
build-backend = "setuptools.build_meta"

[project]
name = "NREL-sup3r"
dynamic = ["version"]
description = "Super Resolving Renewable Resource Data (sup3r)"
keywords = ["sup3r", "NREL"]
readme = "README.rst"
authors = [
  {name = "Brandon Benton", email = "brandon.benton@nrel.gov"},
]
license = {text = "BSD-3-Clause"}
requires-python = ">= 3.8"
classifiers=[
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "NREL-rex>=0.2.84",
  "NREL-phygnn>=0.0.23",
  "NREL-gaps>=0.6.0",
  "NREL-farms>=1.0.4",
  "dask>=2022.0",
  "h5netcdf>=1.1.0",
  "cftime>=1.6.2",
  "matplotlib>=3.1",
  "numpy>=1.7.0",
  "pandas>=2.0",
  "pillow>=10.0",
  "pytest>=5.2",
  "scipy>=1.0.0",
  "sphinx>=7.0",
  "tensorflow>2.4,<2.16",
  "xarray>=2023.0",
]

[project.optional-dependencies]
dev = [
  "build>=0.5",
  "flake8",
  "pre-commit",
  "pylint",
]
doc = [
  "sphinx>=7.0",
  "sphinx_rtd_theme>=2.0",
  "sphinx-click>=4.0",
]
test = [
  "pytest>=5.2",
  "pytest-env"
]

[project.urls]
homepage = "https://github.com/NREL/sup3r"
documentation = "https://nrel.github.io/sup3r/"
repository = "https://github.com/NREL/sup3r"

[tool.ruff]
line-length = 79
indent-width = 4

target-version = "py38"
exclude = [
  "CSV",
  "ref.*?py",
]

[tool.ruff.lint]
fixable = []
# preview = true
# logger-objects = []
task-tags = ["TODO", "FIXME", "XXX"]
select = [
  "A", # flake8-builtins
  "ARG", # flake8-unused-arguments
  "C",
  "C4", # flake8-comprehensions
  "C90", # mccabe
  "COM", # flake8-commas
  "D", # pydocstyle
  "E", # pycodestyle
  "F", # Pyflakes
  "G", # flake8-logging-format
  "I", # isort
  "LOG", # flake8-logging
  "N", # pep8-naming
  "NPY", # numpy-specific
  "PERF", # Perflint
  "PL", # Pylint
  "Q", # flake8-quotes
  "SIM", # flake8-simplify
  "UP", # pyupgrade
  "W", # Warning
  ]

ignore = [
  # Currently don't conform but we might want to reconsider
  "A001", # builtin-variable-shadowing
  # Currently don't conform but we might want to reconsider
  "A002", # builtin-argument-shadowing
  "ARG002", # unused-method-argument
  "ARG003", # unused-class-method-argument
  "ARG004", # unused-static-method-argument
  "ARG005", # unused-lambda-argument
#  "B008", # function-call-in-default-argument
#  "B024", # abstract-base-class-without-abstract-method
#  "B028", # no-explicit-stacklevel
#  "B905", # zip-without-explicit-strict
  "C408", # unnecessary-collection-call
  "C414", # unnecessary-double-cast-or-process
  "COM812", # missing-trailing-comma
  "D105", # undocumented-magic-method
  "D200", # fits-on-one-line
  "D202", # no-blank-line-after-function
  "D204", # one-blank-line-after-class
  "D205", # blank-line-after-summary
  "D207", # under-indentation
  "D209", # new-line-after-last-paragraph
#  "D212", # multi-line-summary-first-line
#  "D213", # multi-line-summary-second-linek
  "D400", # ends-in-period
  "D401", # non-imperative-mood
  "D404", # docstring-starts-with-this
#  "D413", # blank-line-after-last-section
#  "D415", # ends-in-punctuation
#  "E902", # io-error
  "FIX001",	# line-contains-fixme
  # We currently don't conform but we might want to reconsider
  "G001", # logging-string-format
  # We currently don't conform but we might want to reconsider
  "G004", # logging-f-string
  "I001", # unsorted-imports
  "N802", # invalid-function-name
  "N803", # invalid-argument-name
  "N806", # non-lowercase-variable-in-function
  # Consider conforming with NPY002
  "NPY002", # numpy-legacy-random
  "PERF102", # incorrect-dict-iterator
  "PERF203", # try-except-in-loop
  "PERF401", # manual-list-comprehension
  "PLR0904", # too-many-public-methods
  "PLR0912", # too-many-branches
  "PLR0913", # too-many-arguments
  "PLR0914", # too-many-locals
  "PLR0915", # too-many-statements
  "PLR1702", # too-many-nested-blocks
  "PLR1704", # redefined-argument-from-local
  "PLR2004", # magic-value-comparison
  "PLW1514", # unspecified-encoding
  "PLW2901", # redefined-loop-name
  "Q000", # bad-quotes-inline-string
  "Q004", # unnecessary-escaped-quote
  "SIM108", # if-else-block-instead-of-if-exp
  "SIM117", #multiple-with-statements
  "SIM118", # in-dict-keys
  "SIM211", # if-expr-with-false-true
  "UP009", # utf8-encoding-declaration
  "UP015", # redundant-open-modes
  "UP032", # f-string
#  "UP038" # non-pep604-isinstance
  ]
# Ignored in pylint setup but missing on ruff. We shall delete from the
# following lines what is not intended to follow anymore.
# arguments-renamed
# consider-using-f-string
# raw-checker-failed
# bad-inline-option
# locally-disabled
# file-ignored
# suppressed-message
# useless-suppression
# deprecated-pragma
# protected-access
# redefined-outer-name
# redefined-builtin
# broad-except
# logging-format-interpolation
# logging-fstring-interpolation
# wrong-import-order
# wrong-import-position
# relative-beyond-top-level
# too-many-instance-attributes
# too-few-public-methods
# invalid-name
# import-error
# try-except-raise
# no-else-raise
# no-else-return
# unexpected-keyword-arg
# no-value-for-parameter
# too-many-lines
# arguments-differ
# import-outside-toplevel
# super-init-not-called
# isinstance-second-argument-not-valid-type
# inconsistent-return-statements
# no-else-break
# too-many-function-args
# redundant-keyword-arg
# c-extension-no-member

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 5.
max-complexity = 12

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
  "F401", # unused-import
  ]
#"docs/source/conf.py" = [
#  "E402", # unused-import
#  ]

[tool.ruff.lint.pylint]
max-args = 5 # (PLR0913) Maximum number of arguments for function / method
max-bool-expr = 5 # ( PLR0916) Boolean in a single if statement
max-branches=12 # (PLR0912) branches allowed for a function or method body
max-locals=15 # (PLR0912) local variables allowed for a function or method body
max-nested-blocks = 5 # (PLR1702) nested blocks within a function or method body
max-public-methods=20 # (R0904) public methods allowed for a class
max-returns=6 # (PLR0911) return statements for a function or method body
max-statements=50 # (PLR0915) statements allowed for a function or method body

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
# Consider adopting "lf" instead
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.setuptools]
include-package-data = true
packages = ["sup3r"]

[tool.setuptools_scm]
version_file = "sup3r/_version.py"

[tool.pixi.project]
channels = ["conda-forge", "anaconda", "main"]
platforms = ["osx-arm64", "linux-64"]

[tool.pixi.dependencies]
python = "==3.11"
cftime = ">=1.6.2"
dask = ">=2022.0"
h5netcdf = ">=1.1.0"
pillow = ">=10.0"
matplotlib = ">=3.1"
numpy = ">=1.7.0"
pandas = ">=2.0"
scipy = ">=1.0.0"
tensorflow = ">2.4,<2.16"
xarray = ">=2023.0"

[tool.pixi.pypi-dependencies]
NREL-sup3r = { path = ".", editable = true }
NREL-rex = { version = ">=0.2.84" }
NREL-phygnn = { version = ">=0.0.23" }
NREL-gaps = { version = ">=0.6.0" }
NREL-farms = { version = ">=1.0.4" }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev", "doc", "test"], solve-group = "default" }
doc = { features = ["doc"], solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[tool.pixi.tasks]

[tool.pixi.feature.doc.dependencies]
sphinx = ">=7.0"
sphinx_rtd_theme = ">=2.0"

[tool.pixi.feature.test.dependencies]
pytest = ">=5.2"

[tool.pixi.feature.dev.dependencies]
build = ">=0.6"
twine = ">=5.0"
ruff = ">=0.4"

[tool.pytest_env]
CUDA_VISIBLE_DEVICES = "-1"
TF_ENABLE_ONEDNN_OPTS = "0"
